pyproject.toml
[project]
name = "langchain-mcp-adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "isort>=7.0.0",
    "langchain-anthropic>=1.3.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
